backend/requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Chinese NLP
jieba>=0.42.1
pypinyin>=0.50.0
wordfreq>=3.1.0
zhon>=2.0.0  # Chinese character/punctuation detection

# Database
aiosqlite>=0.19.0

# Audio Processing
numpy>=1.24.0
scipy>=1.11.0
# parselmouth-praat>=0.4.3  # Uncomment when needed for Record & Compare
# fastdtw>=0.3.4  # Uncomment when needed for Record & Compare

# TTS (Azure)
azure-cognitiveservices-speech>=1.34.0

# Caching (Optional)
# redis>=5.0.0
# boto3>=1.34.0  # For Cloudflare R2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Development
aiofiles>=23.2.0
